# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavez Prod Backend",
    "description": "Rest Api for Wavez Backend",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    },
    {
      "url": "https://wavez-dev.cf:3000/api"
    },
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
...
    "/stays": {
      "get": {
        "tags": [
          "Stays"
        ],
        "summary": "Get Stays List",
        "operationId": "getStays",
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetStay"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Stays"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Add a new Stays",
        "operationId": "addStays",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateStay"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/stays/{id}": {
      "get": {
        "tags": [
          "Stays"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Get By Vessel Id"
          }
        ],
        "summary": "Get Stay By Id",
        "operationId": "getStayById",
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetStay"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Stays"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Vessel Id"
          },
          {
            "in": "path",
            "name": "vesselType",
            "schema": {
              "type": "string",
              "enum": [
                "STAY"
              ]
            },
            "required": true,
            "description": "Vessel Type {STAY,RENTAL,CHARTER}"
          }
        ],
        "summary": "Update a Stays",
        "operationId": "updateStays",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStay"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateStay"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Stays"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Vessel Id"
          },
          {
            "in": "path",
            "name": "vesselType",
            "schema": {
              "type": "string",
              "enum": [
                "STAY"
              ]
            },
            "required": true,
            "description": "Vessel Type {STAY,RENTAL,CHARTER}"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete a Stays",
        "operationId": "DeleteStays",
        "requestBody": {
          "content": {
          }
        },
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/stays/guest/{id}": {
      "get": {
        "tags": [
          "Guest User"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Get By Vessel Id"
          }
        ],
        "summary": "Get Stay By Id",
        "operationId": "getStayByIdGuest",
        "responses": {
          "200": {
            "description": "Operation Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetStay"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateStay": {
        "type": "object",
        "required": [
          "vesselType"
        ],
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "caption": {
                  "type": "string"
                },
                "imageURL": {
                  "type": "string"
                }
              }
            }
          },
          "numberOfPassengers": {
            "type": "number"
          },
          "vesselLocation": {
            "type": "object",
            "properties": {
              "longitude": {
                "type": "number"
              },
              "latitude": {
                "type": "number"
              }
            }
          },
          "vesselAddress": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string"
              },
              "postalCode": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "countryCode": {
                "type": "string"
              },
              "country": {
                "type": "string"
              },
              "province": {
                "type": "string"
              }
            }
          },
          "rawAddress": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "hasLifeJackets": {
            "type": "boolean"
          },
          "hasFirstAidKit": {
            "type": "boolean"
          },
          "hasFlashlight": {
            "type": "boolean"
          },
          "hasRope": {
            "type": "boolean"
          },
          "vesselType": {
            "type": "string",
            "enum": [
              "STAY"
            ],
            "default": "STAY"
          },
          "vesselCategory": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vesselStatus": {
            "type": "string",
            "enum": [
              "PUBLISHED",
              "UNPUBLISHED"
            ],
            "default": "UNPUBLISHED"
          },
          "vesselYear": {
            "type": "number"
          },
          "numberOfBathrooms": {
            "type": "number"
          },
          "numberOfKitchens": {
            "type": "number"
          },
          "numberOfBeds": {
            "type": "number"
          },
          "vesselPlacement": {
            "type": "string"
          },
          "vesselBrand": {
            "type": "string"
          },
          "vesselModelInfo": {
            "type": "string"
          },
          "vesselLength": {
            "type": "number"
          },
          "serialNumber": {
            "type": "string"
          },
          "vesselWeight": {
            "type": "object",
            "properties": {
              "unit": {
                "type": "string"
              },
              "weight": {
                "type": "number"
              }
            }
          },
          "vesselMaxSpeed": {
            "type": "number"
          },
          "vesselHorsePower": {
            "type": "number"
          },
          "vesselFuelType": {
            "type": "string"
          },
          "vesselPricing": {
            "type": "object",
            "properties": {
              "currency": {
                "type": "string"
              },
              "minimumDeposit": {
                "type": "number"
              },
              "minimumHours": {
                "type": "number"
              },
              "perHour": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perDay": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perWeek": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "vesselFeatures": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "boatLicense": {
            "type": "string"
          },
          "marineSurvey": {
            "type": "string"
          },
          "trailerPlateNumber": {
            "type": "string"
          }
        }
      },
      "CreateStay": {
        "type": "object",
        "required": [
          "title",
          "vesselLocation",
          "vesselAddress",
          "vesselYear",
          "vesselType",
          "vesselCategory",
          "numberOfPassengers",
          "numberOfBathrooms",
          "numberOfKitchens",
          "numberOfBeds"
        ],
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "caption": {
                  "type": "string"
                },
                "imageURL": {
                  "type": "string"
                }
              }
            }
          },
          "numberOfPassengers": {
            "type": "number"
          },
          "vesselLocation": {
            "type": "object",
            "properties": {
              "longitude": {
                "type": "number"
              },
              "latitude": {
                "type": "number"
              }
            }
          },
          "vesselAddress": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string"
              },
              "postalCode": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "countryCode": {
                "type": "string"
              },
              "country": {
                "type": "string"
              },
              "province": {
                "type": "string"
              }
            }
          },
          "rawAddress": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "default": ""
          },
          "vesselBrand": {
            "type": "string"
          },
          "vesselType": {
            "type": "string",
            "enum": [
              "STAY"
            ],
            "default": "STAY"
          },
          "serialNumber": {
            "type": "string"
          },
          "vesselCategory": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vesselStatus": {
            "type": "string",
            "enum": [
              "PUBLISHED",
              "UNPUBLISHED"
            ],
            "default": "UNPUBLISHED"
          },
          "numberOfBathrooms": {
            "type": "number"
          },
          "numberOfKitchens": {
            "type": "number"
          },
          "numberOfBeds": {
            "type": "number"
          },
          "vesselPlacement": {
            "type": "string"
          },
          "vesselModelInfo": {
            "type": "string"
          },
          "vesselLength": {
            "type": "number"
          },
          "vesselWeight": {
            "type": "object",
            "properties": {
              "unit": {
                "type": "string"
              },
              "weight": {
                "type": "number"
              }
            }
          },
          "vesselMaxSpeed": {
            "type": "number"
          },
          "vesselHorsePower": {
            "type": "number"
          },
          "vesselFuelType": {
            "type": "string"
          },
          "vesselPricing": {
            "type": "object",
            "properties": {
              "currency": {
                "type": "string"
              },
              "perHour": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perDay": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perNight": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perWeek": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "vesselFeatures": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "boatLicense": {
            "type": "string"
          },
          "marineSurvey": {
            "type": "string"
          }
        }
      },
...
      "GetStay": {
        "type": "object",
        "properties": {
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "caption": {
                  "type": "string"
                },
                "imageURL": {
                  "type": "string"
                }
              }
            }
          },
          "numberOfPassengers": {
            "type": "number"
          },
          "vesselLocation": {
            "type": "object",
            "properties": {
              "longitude": {
                "type": "number"
              },
              "latitude": {
                "type": "number"
              }
            }
          },
          "vesselAddress": {
            "type": "object",
            "properties": {
              "street": {
                "type": "string"
              },
              "postalCode": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "countryCode": {
                "type": "string"
              },
              "country": {
                "type": "string"
              },
              "province": {
                "type": "string"
              }
            }
          },
          "rawAddress": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "vesselType": {
            "type": "string",
            "enum": [
              "STAY"
            ],
            "default": "STAY"
          },
          "vesselCategory": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vesselStatus": {
            "type": "string",
            "enum": [
              "PUBLISHED",
              "UNPUBLISHED"
            ],
            "default": "UNPUBLISHED"
          },
          "numberOfBathrooms": {
            "type": "number"
          },
          "numberOfKitchens": {
            "type": "number"
          },
          "numberOfBeds": {
            "type": "number"
          },
          "serialNumber": {
            "type": "string"
          },
          "vesselPlacement": {
            "type": "string"
          },
          "vesselYear": {
            "type": "number"
          },
          "vesselBrand": {
            "type": "string"
          },
          "vesselModelInfo": {
            "type": "string"
          },
          "vesselLength": {
            "type": "number"
          },
          "vesselWeight": {
            "type": "object",
            "properties": {
              "unit": {
                "type": "string"
              },
              "weight": {
                "type": "number"
              }
            }
          },
          "vesselMaxSpeed": {
            "type": "number"
          },
          "vesselHorsePower": {
            "type": "number"
          },
          "vesselFuelType": {
            "type": "string"
          },
          "vesselPricing": {
            "type": "object",
            "properties": {
              "currency": {
                "type": "string"
              },
              "perHour": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perDay": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perNight": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              },
              "perWeek": {
                "type": "object",
                "properties": {
                  "checked": {
                    "type": "boolean"
                  },
                  "amount": {
                    "type": "number"
                  }
                }
              }
            }
          },
          "vesselFeatures": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "boatLicense": {
            "type": "string"
          },
          "marineSurvey": {
            "type": "string"
          },
          "trailerPlateNumber": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "BasicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}